requirements.txt
python-dotenv==1.0.1
customtkinter==5.2.2
requests==2.32.3
